+++ libvirt/net/net_linux.go
@@ -9,15 +9,18 @@ import (
 	"errors"
 	"fmt"
 	stdnet "net"
+	"slices"
 	"strconv"
 	"time"
 
 	"github.com/coreos/go-iptables/iptables"
 	"github.com/hashicorp/go-multierror"
+	"github.com/hashicorp/go-set"
 	"github.com/hashicorp/nomad-driver-virt/libvirt"
 	"github.com/hashicorp/nomad-driver-virt/virt/net"
 	"github.com/hashicorp/nomad/plugins/drivers"
 	"github.com/hashicorp/nomad/plugins/shared/structs"
+	lv "libvirt.org/go/libvirt"
 )
 
 const (
@@ -168,7 +171,6 @@ func ensureIPTablesChain(ipt *iptables.IPTables, table, chain string) (bool, err
 }
 
 func (c *Controller) VMStartedBuild(req *net.VMStartedBuildRequest) (*net.VMStartedBuildResponse, error) {
-
 	if req == nil {
 		return nil, errors.New("net controller: no request provided")
 	}
@@ -199,7 +201,7 @@ func (c *Controller) VMStartedBuild(req *net.VMStartedBuildRequest) (*net.VMStar
 		return nil, fmt.Errorf("failed to lookup network: %w", err)
 	}
 
-	ipAddr, err := c.discoverDHCPLeaseIP(network, req.DomainName, networkName)
+	ipAddr, err := c.discoverDHCPLeaseIP(network, req.Hostname, networkName, req.Hwaddrs)
 	if err != nil {
 		return nil, fmt.Errorf("failed to discover IP address: %w", err)
 	}
@@ -253,7 +255,7 @@ func (c *Controller) networkNameFromBridgeName(name string) (string, error) {
 // network. The function includes a ticker in order to poll for the information
 // as this can take several seconds to become available.
 func (c *Controller) discoverDHCPLeaseIP(
-	network libvirt.ConnectNetworkShim, domainName, netName string) (string, error) {
+	network libvirt.ConnectNetworkShim, hostname, netName string, hwaddrs []string) (string, error) {
 
 	ticker := time.NewTicker(c.dhcpLeaseDiscoveryInterval)
 	defer ticker.Stop()
@@ -261,6 +263,7 @@ func (c *Controller) discoverDHCPLeaseIP(
 	timeout := time.NewTimer(c.dhcpLeaseDiscoveryTimeout)
 	defer timeout.Stop()
 
+	macs := set.From(hwaddrs)
 	for {
 		select {
 		case <-ticker.C:
@@ -270,7 +273,8 @@ func (c *Controller) discoverDHCPLeaseIP(
 			// debug entry while performing this "long-lived" process should
 			// help operators understand what is happening.
 			c.logger.Debug("attempting DHCP lease discovery",
-				"domain", domainName, "network_name", netName)
+				"hostname", hostname, "network_name", netName,
+				"hwaddrs", hwaddrs)
 
 			// Lookup the DHCP leases of the network. If we receive any error,
 			// log and try again. If it is transient error, we will find the
@@ -283,14 +287,47 @@ func (c *Controller) discoverDHCPLeaseIP(
 				continue
 			}
 
+			matches := []lv.NetworkDHCPLease{}
+
+			// Gather all matching leases
 			for _, lease := range dhcpLeases {
-				if lease.Hostname == domainName {
-					return lease.IPaddr, nil
+				// Check if lease matches any available interfaces on the domain.
+				if !macs.Contains(lease.Mac) {
+					continue
 				}
+
+				// Check if the hostname is set, and matches
+				if lease.Hostname != "" && lease.Hostname != hostname {
+					continue
+				}
+
+				// Only want to add leases that are still valid.
+				if lease.ExpiryTime.Before(time.Now()) {
+					continue
+				}
+
+				c.logger.Debug("DHCP lease detected", "hostname", hostname, "network_name", netName,
+					"hwaddrs", hwaddrs, "lease", lease)
+				matches = append(matches, lease)
 			}
 
+			if len(matches) == 0 {
+				continue
+			}
+
+			// If any matches were found, sort them in descending order
+			// by the lease expiry date and return the address with the
+			// expiry furthest in the future. This is done to handle situations
+			// where an interface's MAC address is being set and the instance
+			// has been destroyed and created again resulting in multiple
+			// leases for the same MAC.
+			slices.SortFunc(matches, func(a, b lv.NetworkDHCPLease) int {
+				return b.ExpiryTime.Compare(a.ExpiryTime)
+			})
+
+			return matches[0].IPaddr, nil
 		case <-timeout.C:
-			return "", fmt.Errorf("timeout reached discovering DHCP lease for %q", domainName)
+			return "", fmt.Errorf("timeout reached discovering DHCP lease for %q", hostname)
 		}
 	}
 }