+++ virt/net/net.go
@@ -27,8 +27,10 @@ const (
 // sub-system to perform its configuration, once a VM has been started.
 type VMStartedBuildRequest struct {
 	DomainName string
+	Hostname   string
 	NetConfig  *NetworkInterfacesConfig
 	Resources  *drivers.Resources
+	Hwaddrs    []string
 }
 
 // VMStartedBuildResponse is the response sent object once the network