+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## UNRELEASED
 
+* build: Update Nomad verison to 1.10.0 [GH-111](https://github.com/hashicorp/nomad-driver-virt/pull/111)
+* build: Update Go to 1.24.2 [GH-111](https://github.com/hashicorp/nomad-driver-virt/pull/111)
+
 ## 0.0.1-beta.1 (15 October, 2024)
 
 First beta release of initial v0.1.0 version.