+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## UNRELEASED
 
+BUG FIXES:
+
+* libvirt: Automatically reconnect when required [GH-129](https://github.com/hashicorp/nomad-driver-virt/pull/129)
+
+IMPROVEMENTS:
+
 * build: Update Nomad verison to 1.10.0 [GH-111](https://github.com/hashicorp/nomad-driver-virt/pull/111)
 * build: Update Go to 1.24.2 [GH-111](https://github.com/hashicorp/nomad-driver-virt/pull/111)