+++ CHANGELOG.md
@@ -8,6 +8,7 @@ IMPROVEMENTS:
 
 * build: Update Nomad verison to 1.10.0 [GH-111](https://github.com/hashicorp/nomad-driver-virt/pull/111)
 * build: Update Go to 1.24.2 [GH-111](https://github.com/hashicorp/nomad-driver-virt/pull/111)
+* net: Perform DHCP lookup using MAC address [GH-131](https://github.com/hashicorp/nomad-driver-virt/pull/131)
 
 ## 0.0.1-beta.1 (15 October, 2024)